MusicPlayer/requirements.txt
### requirements.txt

# Paste
# pip install -r requirements.txt
# to install the dependencies

PySide6
yt-dlp
mutagen
requests
python-vlc
keyboard
pynput